restructure to match separate-v2.1 branch
